Summarise Startle

Returns a mathematical summary of the startle data. Returned list with two data frames. First contains calculated median, interquartile range, mean and standard deviation for values pertaining to the period of the pulse and in between pulses. The second contains the results of the desired statistical test, by default Wilcox is used. User can specify which test will be used with method argument.

summariseStartle(df, method = "wilcox.test")

Arguments

df
A dataframe returned by loadStartleData function.

method
A parameter defining what type of statistical testing will be used.

Value

A mathematical summary in a list.

Examples

# Load example data
df <- ratpasta_demo

# running examples
summariseStartle(df)

#> $Values
#> # A tibble: 8 x 9
#> # Groups:   stage, impulse [4]
#>   stage impulse lgroup `mean(value)` `sd(value)` `min(value)` `median(value)`
#>   <fct> <fct>   <chr>          <dbl>       <dbl>        <dbl>           <dbl>
#> 1 W/O … No I    ctr             3.25        3.46    0.00195              2.27
#> 2 W/O … No I    stz             6.37       13.4     0.0000971            2.41
#> 3 W/O … I       ctr            11.4        23.1     0.00195              3.84
#> 4 W/O … I       stz            62.1       112.      0.00255             26.3 
#> 5 W PI  No I    ctr             4.88        5.37    0.00195              3.06
#> 6 W PI  No I    stz             2.92        3.43    0.0000971            1.86
#> 7 W PI  I       ctr             7.99       10.2     0.00195              4.48
#> 8 W PI  I       stz             8.94       14.5     0.0000971            3.44
#> # … with 2 more variables: `IQR(value)` <dbl>, `max(value)` <dbl>
#> 
#> $Ratio
#> # A tibble: 4 x 8
#> # Groups:   stage [2]
#>   stage lgroup `mean(ratio)` `sd(ratio)` `min(ratio)` `median(ratio)`
#>   <fct> <chr>          <dbl>       <dbl>        <dbl>           <dbl>
#> 1 W/O … ctr             1.83       0.897        0.885            1.48
#> 2 W/O … stz            11.4       12.5          0.607            4.05
#> 3 W PI  ctr             1.61       0.972        0.667            1.26
#> 4 W PI  stz             1.95       0.682        0.756            1.86
#> # … with 2 more variables: `IQR(ratio)` <dbl>, `max(ratio)` <dbl>
#> 
#> $RatioS
#> # A tibble: 6 x 8
#>   .y.   group1     group2           p p.adj p.format p.signif method  
#>   <chr> <chr>      <chr>        <dbl> <dbl> <chr>    <chr>    <chr>   
#> 1 ratio ctr W/O PP stz W/O PP 0.0155  0.077 0.0155   *        Wilcoxon
#> 2 ratio ctr W/O PP ctr W PP   0.211   0.35  0.2110   ns       Wilcoxon
#> 3 ratio ctr W/O PP stz W PP   0.174   0.35  0.1738   ns       Wilcoxon
#> 4 ratio stz W/O PP ctr W PP   0.00427 0.026 0.0043   **       Wilcoxon
#> 5 ratio stz W/O PP stz W PP   0.0460  0.14  0.0460   *        Wilcoxon
#> 6 ratio ctr W PP   stz W PP   0.0181  0.077 0.0181   *        Wilcoxon
#>